Try to fix the conversion implementation

+++ src/vm/types/cp_types/vm_double.c
@@ -1,7 +1,5 @@
 #include "vm_double.h"
 
-#include <math.h>
-
 double vm_itod(uint32_t low, uint32_t high) {
-    return vm_itolf(low, high);
+    return (double)vm_itolf(low, high);
 }
